requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
django-shapefileimport==1.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
Fiona==1.9.4.post1
GDAL==3.7.2
GeoAlchemy2==0.14.1
geopandas==0.14.0
greenlet==2.0.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
openapi-codec==1.3.2
packaging==23.1
pandas==2.1.1
psycopg2-binary==2.9.7
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
shapely==2.0.1
simplejson==3.19.2
six==1.16.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
